// src/app/shared/interfaces/ievent.ts
export interface IEvent {
   
        creatorID: string;             
        currentAttendees: string;      
        date: string;                 
        description: string;         
        direccion: string;             
        duration: number;              
        eventID: string;               
        numberOfAttendees: number;     
        specialRequirements: string;   
        typeEvent: string;             
}